edit reaction scheme, introduce reaction_radius
The reaction scheme is now as in single_dimer_pbc: the evaluation of the
reaction probability is done in change_species and not in flag_particles.
The change_species routine relies first on the md_flag and then checks for the
reaction radius distance that is set by the user.